+++ src/ast/transform-parent.ts
@@ -4,7 +4,14 @@ import { ProjectId } from "../model/project-id.ts";
 import { NextIdentifierNumberGetter } from "../model/task-id-number.ts";
 import { transformNode } from "./transform-node.ts";
 
-export function replaceParent<
+/**
+ * Transforms the given {@link Parent}.
+ * @param projectId The project identifier to use.
+ * @param nextIdentifierNumberGetter The function to use to get the next identifier number.
+ * @param parent The parent node to transform.
+ * @returns The transformed parent node.
+ */
+export function transformParent<
   T extends Parent,
   PI extends ProjectId = ProjectId,
 >(
