+++ src/markdown/update-markdown-link-node.ts
@@ -1,5 +1,5 @@
 import { Link } from "npm:@types/mdast";
-import { isText } from "../ast/types.ts";
+import { isText } from "../ast/node-types.ts";
 import { sequence } from "../strings/regex.ts";
 import { basename } from "std/path/basename.ts";
 import { extname } from "std/path/extname.ts";
