+++ src/model/box.ts
@@ -1,11 +1,17 @@
+import { Heading, ListItem, PhrasingContent } from "npm:@types/mdast";
+import { isHeading, isListItem, isText } from "../ast/node-types.ts";
 import { capture, sequence } from "../strings/regex.ts";
+import { StringStartingWith } from "../strings/string-types.ts";
 import {
   isOnlyA,
   startsWithA,
   TextTypeGuard,
 } from "../strings/text-type-guard.ts";
 import { PROJECT_ID_REGEX, ProjectId } from "./project-id.ts";
-import { createTaskIdPlaceholderRegex } from "./task-id-placeholder.ts";
+import {
+  createTaskIdPlaceholderRegex,
+  TaskIdPlaceholder,
+} from "./task-id-placeholder.ts";
 import { createTaskIdRegex, TaskId } from "./task-id.ts";
 
 export type BoxContents = " " | "x";
@@ -21,7 +27,11 @@ export const BOX_REGEX = capture(
 );
 export const isBox: TextTypeGuard<Box> = isOnlyA<Box>(BOX_REGEX);
 
-export type BoxAndTaskId = `${Box} ${TaskId}`;
+export type BoxAndTaskId<PI extends ProjectId = ProjectId> = `${Box} ${TaskId<
+  PI
+>}`;
+export type BoxAndTaskIdPlaceholder<PI extends ProjectId = ProjectId> =
+  `${Box} ${TaskIdPlaceholder<PI>}`;
 export function createBoxAndTaskIdRegex<PI extends ProjectId = ProjectId>(
   projectId: PI | RegExp = PROJECT_ID_REGEX,
 ) {
@@ -35,18 +45,82 @@ export function createBoxAndTaskIdPlaceholderRegex<
   return sequence(BOX_REGEX, " ", createTaskIdPlaceholderRegex(projectId));
 }
 
-export function createIsBoxAndTaskId(
-  projectId: ProjectId | RegExp = PROJECT_ID_REGEX,
-): TextTypeGuard<BoxAndTaskId> {
-  return isOnlyA<BoxAndTaskId>(
+export function createIsBoxAndTaskId<PI extends ProjectId = ProjectId>(
+  projectId: PI | RegExp = PROJECT_ID_REGEX,
+): TextTypeGuard<BoxAndTaskId<PI>> {
+  return isOnlyA<BoxAndTaskId<PI>>(
     createBoxAndTaskIdRegex(projectId),
   );
 }
 
-export function createExtractBoxAndTaskId(
-  projectId: ProjectId | RegExp = PROJECT_ID_REGEX,
-): TextTypeGuard<BoxAndTaskId> {
+export function createExtractBoxAndTaskId<PI extends ProjectId = ProjectId>(
+  projectId: PI | RegExp = PROJECT_ID_REGEX,
+): TextTypeGuard<BoxAndTaskId<PI>> {
   return startsWithA<
-    BoxAndTaskId
+    BoxAndTaskId<PI>
   >(createBoxAndTaskIdRegex(projectId));
+}
+
+/**
+ * A {@link TextTypeGuard} that checks if a string starts with a box.
+ */
+export const startsWithBox: TextTypeGuard<StringStartingWith<Box>> =
+  startsWithA(BOX_REGEX);
+
+/**
+ * Returns true if the given {@link ListItem} has a box.
+ * @param listItem The list item to check.
+ */
+export function hasBox(
+  listItem: ListItem,
+): listItem is ListItem & { checked: boolean };
+/**
+ * Returns true if the given {@link Heading} has a box in its first child.
+ * @param heading The heading to check.
+ */
+export function hasBox(
+  heading: Heading,
+): heading is Heading & {
+  children: [
+    { type: "text"; value: StringStartingWith<Box> },
+    ...PhrasingContent[],
+  ];
+};
+/**
+ * Returns true if the given {@link ListItem} or {@link Heading} has a box.
+ * @param node The node to check.
+ */
+export function hasBox<T extends ListItem | Heading>(node: T): boolean {
+  if (isHeading(node)) {
+    return hasHeadingBox(node);
+  }
+  if (isListItem(node)) {
+    return hasListItemBox(node);
+  }
+  return false;
+}
+
+/**
+ * Type-guard for {@link ListItem} that definitely has a box.
+ * @param listItem The list item to check.
+ */
+function hasListItemBox(
+  listItem: ListItem,
+): listItem is ListItem & { checked: boolean } {
+  return listItem.checked === true || listItem.checked === false;
+}
+
+/**
+ * Returns true if the given heading has a first child that is a text node, and that starts with a box.
+ * @param heading The heading to check.
+ */
+function hasHeadingBox(
+  heading: Heading,
+): heading is Heading & {
+  children: [
+    { type: "text"; value: StringStartingWith<Box> },
+    ...PhrasingContent[],
+  ];
+} {
+  return isText(heading.children[0]) && startsWithBox(heading.children[0]);
 }
